+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 Release Notes
 ====
 
+# 01-12-2020
+<a href="https://www.nuget.org/packages/dotnext.net.cluster/1.2.11">DotNext.Net.Cluster 1.2.11</a>
+* Ability to reconstruct internal state using `PersistentState.ReplayAsync` method
+
+# 01-11-2020
+<a href="https://www.nuget.org/packages/dotnext/1.2.10">DotNext 1.2.10</a>
+* Fixed invalid behavior of `StreamSegment.Position` property
+
+<a href="https://www.nuget.org/packages/dotnext.net.cluster/1.2.10">DotNext.Net.Cluster 1.2.10</a>
+* Removed redundant validation of log entry index in `PersistentState`
+
+# 12-06-2019
+<a href="https://www.nuget.org/packages/dotnext.unsafe/1.2.10">DotNext.Unsafe 1.2.10</a>
+* Fixed invalid usage of `GC.RemoveMemoryPressure` in `Reallocate` methods
+
 # 12-04-2019
 <a href="https://www.nuget.org/packages/dotnext/1.2.9">DotNext 1.2.9</a>
 * `UserDataStorage` no longer stores **null** values in its internal dictionary